Refactor all types modules and classes in common
