+++ packages/client/src/images/maps/index.ts
@@ -1,9 +1,7 @@
 import dungeon from './dungeon.png';
 
 // This is where you want to reference images needed by your maps
-const SpriteSheets: {
-  [key: string]: string;
-} = {
+const SpriteSheets: { [key: string]: string } = {
   'dungeon.png': dungeon,
 };
